Move ConnectionOverNetty and StreamOverNetty out of io.libp2p.core
The TCPTransport is built over netty, so these two classes belong there.
The Connection and Stream interfaces are better, but still need further
filling out.  Specifically, I think a Connection should create a Stream,
as in the other implementations.

These two classes have deliberately awkward names to help flag up where
in the codebase they're used.
