+++ src/main/kotlin/io/libp2p/core/Stream.kt
@@ -1,9 +1,10 @@
 package io.libp2p.core
 
-import io.libp2p.etc.PROTOCOL
-import io.netty.channel.Channel
 import java.util.concurrent.CompletableFuture
 
+/**
+ * Represents a multiplexed stream over wire connection
+ */
 interface Stream : P2PChannel {
     val connection: Connection
 
@@ -19,31 +20,4 @@ interface Stream : P2PChannel {
     fun getProtocol(): CompletableFuture<String>
 
     fun writeAndFlush(msg: Any)
-}
-
-/**
- * Represents a multiplexed stream over wire connection
- */
-class StreamOverNetty(
-    ch: Channel,
-    override val connection: Connection
-) : Stream, P2PAbstractChannel(ch) {
-    init {
-        nettyChannel.attr(PROTOCOL).set(CompletableFuture())
-    }
-
-    /**
-     * Returns the [PeerId] of the remote peer [Connection] which this
-     * [Stream] created on
-     */
-    override fun remotePeerId() = connection.secureSession.remoteId
-
-    /**
-     * @return negotiated protocol
-     */
-    override fun getProtocol(): CompletableFuture<String> = nettyChannel.attr(PROTOCOL).get()
-
-    override fun writeAndFlush(msg: Any) {
-        nettyChannel.writeAndFlush(msg)
-    }
 }
